+++ src/form/lean.hpp
@@ -35,6 +35,8 @@ class LeanFormula {
   bool convertToLean(Expression& expr, Number patternOffset,
                      bool insideOfLocalFunc);
 
+  bool needsIntToNat(const Expression& expr) const;
+
   bool isLocalOrSeqFunc(const std::string& funcName) const;
 
   std::string printFunction(const std::string& funcName) const;